consolidate durableUtils, add connectionSetting validate files
