+++ src/commands/appSettings/connectionSettings/eventHubs/validateEventHubsConnection.ts
@@ -1,0 +1,71 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { AzExtFsExtra, AzureWizard, AzureWizardExecuteStep, AzureWizardPromptStep } from "@microsoft/vscode-azext-utils";
+import * as path from "path";
+import { CodeAction, ConnectionKey, hostFileName, localEventHubsEmulatorConnectionRegExp } from "../../../../constants";
+import { IHostJsonV2, INetheriteTaskJson } from "../../../../funcConfig/host";
+import { getLocalSettingsConnectionString } from "../../../../funcConfig/local.settings";
+import { NetheriteConfigureHostStep } from "../../../createFunction/durableSteps/netherite/NetheriteConfigureHostStep";
+import { NetheriteEventHubNameStep } from "../../../createFunction/durableSteps/netherite/NetheriteEventHubNameStep";
+import { IConnectionPromptOptions } from "../IConnectionPromptOptions";
+import { ISetConnectionSettingContext } from "../ISetConnectionSettingContext";
+import { EventHubsConnectionExecuteStep } from "./EventHubsConnectionExecuteStep";
+import { EventHubsConnectionPromptStep } from "./EventHubsConnectionPromptStep";
+import { IEventHubsConnectionWizardContext } from "./IEventHubsConnectionWizardContext";
+
+// Supports validation on both 'debug' and 'deploy'
+export async function validateEventHubsConnection(context: Omit<ISetConnectionSettingContext, 'projectPath'>, projectPath: string, options?: IConnectionPromptOptions): Promise<void> {
+    const eventHubsConnection: string | undefined = await getLocalSettingsConnectionString(context, ConnectionKey.EventHubs, projectPath);
+    const eventHubName: string | undefined = await getEventHubName(projectPath);
+
+    if (!!eventHubsConnection && !!eventHubName) {
+        if (context.action === CodeAction.Deploy) {
+            if (!localEventHubsEmulatorConnectionRegExp.test(eventHubsConnection)) {
+                // Found a valid connection in deploy mode. Set it and skip the wizard.
+                context[ConnectionKey.EventHubs] = eventHubsConnection;
+                return;
+            }
+            // Found an invalid connection for deploy mode, we need to proceed with acquiring a connection through the wizard...
+        } else {
+            // Found a valid connection in debug mode.  Skip the wizard.
+            return;
+        }
+    }
+
+    const wizardContext: IEventHubsConnectionWizardContext = Object.assign(context, { projectPath });
+    const promptSteps: AzureWizardPromptStep<IEventHubsConnectionWizardContext>[] = [];
+    const executeSteps: AzureWizardExecuteStep<IEventHubsConnectionWizardContext>[] = [];
+
+    if (!eventHubsConnection) {
+        promptSteps.push(new EventHubsConnectionPromptStep(options));
+        executeSteps.push(new EventHubsConnectionExecuteStep());
+    }
+
+    if (!eventHubName) {
+        promptSteps.push(new NetheriteEventHubNameStep());
+    }
+
+    executeSteps.push(new NetheriteConfigureHostStep());
+
+    const wizard: AzureWizard<IEventHubsConnectionWizardContext> = new AzureWizard(wizardContext, {
+        promptSteps,
+        executeSteps
+    });
+
+    await wizard.prompt();
+    await wizard.execute();
+}
+
+export async function getEventHubName(projectPath: string): Promise<string | undefined> {
+    const hostJsonPath = path.join(projectPath, hostFileName);
+    if (!await AzExtFsExtra.pathExists(hostJsonPath)) {
+        return undefined;
+    }
+
+    const hostJson: IHostJsonV2 = await AzExtFsExtra.readJSON(hostJsonPath);
+    const taskJson: INetheriteTaskJson = hostJson.extensions?.durableTask as INetheriteTaskJson;
+    return taskJson?.hubName;
+}
